API reference

Only exported names are part of the API.

MultiAgentPathFinding — Module

MultiAgentPathFinding

A package for Multi-Agent Path Finding instances and algorithms.

Exports

EdgeConflict

MAPF

Reservation

Solution

VertexConflict

cooperative_astar

find_conflict

independent_dijkstra

is_feasible

list_map_names

list_scenario_names

nb_agents

path_cost

select_agents

solution_cost

Structures

MultiAgentPathFinding.MAPF — Type

struct MAPF{W, VC, EC}

Instance of a Multi-Agent Path Finding problem with custom conflict rules.

Constructors

MAPF(
    g::AbstractGraph, departures::Vector{Int}, arrivals::Vector{Int};
    vertex_conflicts=LazyVertexConflicts(), edge_conflicts=LazyEdgeConflicts()
)

Fields

g::SimpleWeightedGraphs.SimpleWeightedGraph{Int64, W} where W: underlying weighted graph

departures::Vector{Int64}: agent departure vertices

arrivals::Vector{Int64}: agent arrival vertices

vertex_conflicts::Any: indexable object linking vertices to their incompatibility set

edge_conflicts::Any: indexable object linking edges (as tuples) to their incompatibility set

vertex_to_coord::Union{Missing, Vector{Tuple{Int64, Int64}}}: mapping from integer vertices to coordinate tuples

MultiAgentPathFinding.Solution — Type

struct Solution

Store one path for each agent of a MAPF.

Fields

paths::Vector{Vector{Int64}}

MultiAgentPathFinding.Reservation — Type

struct Reservation

Keep track of which vertices and edges are known to be occupied and by whom.

It does not have to be a physical occupation: some agent might be occupying a related vertex or edge which generates a conflict.

Fields

single_occupied_vertices::Dict{Tuple{Int64, Int64}, Int64}: (t, v) -> a where a is the only agent occupying v at time t

single_occupied_edges::Dict{Tuple{Int64, Int64, Int64}, Int64}: (t, u, v) -> a where a is the only agent occupying (u, v) at time t

multi_occupied_vertices::Dict{Tuple{Int64, Int64}, Vector{Int64}}: (t, v) -> [a1, a2] where a1, a2 are the multiple agents occupying v at time t

multi_occupied_edges::Dict{Tuple{Int64, Int64, Int64}, Vector{Int64}}: (t, u, v) -> [a1, a2] where a1, a2 are the multiple agents occupying (u, v) at time t

arrival_vertices::Dict{Int64, Tuple{Int64, Int64}}: v -> (t, a) where a is the agent whose arrival vertex is v and who owns it starting at time t (necessary for stay-at-target behavior)

arrival_vertices_crossings::Dict{Int64, Vector{Tuple{Int64, Int64}}}: v -> [(t2, a1), (t2, a2)] where the ai are additional agents visiting vertex v at times ti, once it is already owned as an arrival vertex

Note

The split between single and multi is done for efficiency reasons: there will be many more single_occupied than multi_occupied, so allocating a set for all of these would be wasteful (and using Union{Int, Set{Int}} would be type-unstable).

Access

MultiAgentPathFinding.nb_agents — Function

nb_agents(mapf::MAPF) -> Int64

Count the number of agents in mapf.

MultiAgentPathFinding.select_agents — Function

select_agents(mapf::MAPF, agents::AbstractVector{<:Integer})

Select a subset of agents in mapf and return a new MAPF.

Feasibility and cost

MultiAgentPathFinding.path_cost — Function

path_cost(
    path::Vector{Int64},
    g::SimpleWeightedGraphs.SimpleWeightedGraph
) -> Any

Sum the costs of all the edges in path within mapf.

MultiAgentPathFinding.solution_cost — Function

solution_cost(solution::Solution, mapf::MAPF) -> Any

Sum the costs of all the paths in solution. Costs are computed within mapf for each agent.

MultiAgentPathFinding.is_feasible — Function

is_feasible(solution::Solution, mapf::MAPF; verbose) -> Bool

Check whether solution is both individually and collectively feasible (correct paths and no conflicts).

MultiAgentPathFinding.find_conflict — Function

find_conflict(
    solution::Solution,
    mapf::MAPF
) -> Union{Nothing, EdgeConflict, VertexConflict}

Find a conflict in solution for mapf.

MultiAgentPathFinding.VertexConflict — Type

struct VertexConflict

Temporal vertex conflict between two agents (for debugging purposes).

Fields

t::Int64: time

v::Int64: vertex

a1::Int64: first agent

a2::Int64: second agent

MultiAgentPathFinding.EdgeConflict — Type

struct EdgeConflict

Temporal edge conflict between two agents (for debugging purposes).

Fields

t::Int64: time

u::Int64: edge source

v::Int64: edge destination

a1::Int64: first agent

a2::Int64: second agent

Basic algorithms

MultiAgentPathFinding.independent_dijkstra — Function

independent_dijkstra(mapf::MAPF) -> Solution

Compute independent shortest paths for each agent of mapf.

Returns a Solution where some paths may be empty if the vertices are not connected.

MultiAgentPathFinding.cooperative_astar — Function

cooperative_astar(mapf::MAPF) -> Solution
cooperative_astar(
    mapf::MAPF,
    agents::AbstractVector{<:Integer}
) -> Solution

Solve a MAPF problem mapf for a set of agents with the cooperative A* algorithm of Silver (2005), see https://ojs.aaai.org/index.php/AIIDE/article/view/18726.

Returns a Solution where some paths may be empty if the vertices are not connected.

Benchmarks

MultiAgentPathFinding.list_map_names — Function

list_map_names() -> Vector{String}

List available maps from the benchmark set.

MultiAgentPathFinding.list_scenario_names — Function

list_scenario_names(scen_type::String) -> Vector{String}

List available scenarios from the benchmark set.

Internals

MultiAgentPathFinding.LazySwappingConflicts — Type

struct LazySwappingConflicts

Lazy dict-like storage for the mapping (u, v) -> [(v, u)].

MultiAgentPathFinding.LazyVertexConflicts — Type

struct LazyVertexConflicts

Lazy dict-like storage for the mapping v -> [v].

MultiAgentPathFinding.MAPFBenchmarkAgent — Type

struct MAPFBenchmarkAgent

Encode one agent of a MAPF scenario.

Fields

index::Int64

bucket::Int64

map_path::String

width::Int64

height::Int64

start_i::Int64

start_j::Int64

goal_i::Int64

goal_j::Int64

optimal_length::Float64

MultiAgentPathFinding.arrive! — Method

arrive!(
    reservation::Reservation,
    a::Integer,
    t::Integer,
    v::Integer
)

Update reservation so that agent a arrives vertex v at time t and never moves again.

MultiAgentPathFinding.cell_color — Method

cell_color(
    c::Char
) -> ColorTypes.RGB{FixedPointNumbers.N0f8}

Give a color object corresponding to the type of cell.

To visualize a map in VSCode, just run cell_color.(map_matrix) in the REPL.

MultiAgentPathFinding.exists_in_graph — Method

exists_in_graph(
    path::Vector{Int64},
    g::SimpleWeightedGraphs.SimpleWeightedGraph
) -> Bool

Check that path is feasible in the graph g, i.e. that all vertices and edges exist.

MultiAgentPathFinding.is_collectively_feasible — Method

is_collectively_feasible(
    solution::Solution,
    mapf::MAPF;
    verbose
) -> Bool

Check whether solution contains any conflicts between agents.

MultiAgentPathFinding.is_individually_feasible — Method

is_individually_feasible(
    solution::Solution,
    mapf::MAPF;
    verbose
) -> Bool

Check whether solution is feasible when agents are considered separately.

MultiAgentPathFinding.is_occupied_edge — Method

is_occupied_edge(
    reservation::Reservation,
    t::Integer,
    u::Integer,
    v::Integer
) -> Bool

Check whether edge (u, v) is occupied at time t in a reservation.

MultiAgentPathFinding.is_occupied_vertex — Method

is_occupied_vertex(
    reservation::Reservation,
    t::Integer,
    v::Integer
) -> Any

Check whether vertex v is occupied at time t in a reservation.

MultiAgentPathFinding.map_from_scenario — Method

map_from_scenario(scenario_name::AbstractString) -> String

Return the map associated with a benchmark scenario.

MultiAgentPathFinding.occupy! — Method

occupy!(
    reservation::Reservation,
    a::Integer,
    t::Integer,
    v::Integer
)

Update reservation so that agent a occupies vertex v at time t.

MultiAgentPathFinding.occupy! — Method

occupy!(
    reservation::Reservation,
    a::Integer,
    t::Integer,
    u::Integer,
    v::Integer
)

Update reservation so that agent a occupies edge (u, v) at time t.

MultiAgentPathFinding.parse_benchmark_map — Method

parse_benchmark_map(
    map_matrix::AbstractMatrix
) -> Tuple{SimpleWeightedGraphs.SimpleWeightedGraph{Int64, Float64}, Dict{Tuple{Int64, Int64}, Int64}, Vector{Tuple{Int64, Int64}}}

Create a sparse grid graph from a map specified as a matrix of characters.

MultiAgentPathFinding.parse_benchmark_scenario — Method

parse_benchmark_scenario(
    scenario::Vector{MultiAgentPathFinding.MAPFBenchmarkAgent}
) -> Tuple{Vector{Tuple{Int64, Int64}}, Vector{Tuple{Int64, Int64}}}

Turn a scenario into vectors of departure coordinates and a vector of arrival coordinates.

MultiAgentPathFinding.read_benchmark_map — Method

read_benchmark_map(map_name::AbstractString) -> Matrix{Char}

Read a map matrix from a text file.

Returns a Matrix{Char}.

MultiAgentPathFinding.read_benchmark_scenario — Method

read_benchmark_scenario(
    scenario_name::AbstractString,
    map_name::AbstractString
) -> Vector{MultiAgentPathFinding.MAPFBenchmarkAgent}

Read a scenario from a text file, and check that it corresponds to a given map.

Returns a Vector{MAPFBenchmarkAgent}.

MultiAgentPathFinding.scenarios_from_map — Method

scenarios_from_map(
    map_name::AbstractString,
    scen_type::String
) -> Vector{String}

List the scenarios associated with a benchmark map.

MultiAgentPathFinding.update_reservation! — Method

update_reservation!(
    reservation::Reservation,
    path::Vector{Int64},
    a::Integer,
    mapf::MAPF
)

Add the vertices and edges occupied by a path to a reservation.

Index

MultiAgentPathFinding

MultiAgentPathFinding.EdgeConflict

MultiAgentPathFinding.LazySwappingConflicts

MultiAgentPathFinding.LazyVertexConflicts

MultiAgentPathFinding.MAPF

MultiAgentPathFinding.MAPFBenchmarkAgent

MultiAgentPathFinding.Reservation

MultiAgentPathFinding.Solution

MultiAgentPathFinding.VertexConflict

MultiAgentPathFinding.arrive!

MultiAgentPathFinding.cell_color

MultiAgentPathFinding.cooperative_astar

MultiAgentPathFinding.exists_in_graph

MultiAgentPathFinding.find_conflict

MultiAgentPathFinding.independent_dijkstra

MultiAgentPathFinding.is_collectively_feasible

MultiAgentPathFinding.is_feasible

MultiAgentPathFinding.is_individually_feasible

MultiAgentPathFinding.is_occupied_edge

MultiAgentPathFinding.is_occupied_vertex

MultiAgentPathFinding.list_map_names

MultiAgentPathFinding.list_scenario_names

MultiAgentPathFinding.map_from_scenario

MultiAgentPathFinding.nb_agents

MultiAgentPathFinding.occupy!

MultiAgentPathFinding.occupy!

MultiAgentPathFinding.parse_benchmark_map

MultiAgentPathFinding.parse_benchmark_scenario

MultiAgentPathFinding.path_cost

MultiAgentPathFinding.read_benchmark_map

MultiAgentPathFinding.read_benchmark_scenario

MultiAgentPathFinding.scenarios_from_map

MultiAgentPathFinding.select_agents

MultiAgentPathFinding.solution_cost

MultiAgentPathFinding.update_reservation!