API reference

Docs

MultiAgentPathFinding.MultiAgentPathFinding — Module

MultiAgentPathFinding

A package for Multi-Agent Path Finding instances and algorithms.

MultiAgentPathFinding.Conflict — Type

struct Conflict

Store a conflict between two agents for debugging purposes.

Fields

name::Symbol: type of conflict (:vertex or :vdge)

a1::Int64: first agent

a2::Int64: second agent

t1::Int64: time for the first agent

t2::Int64: time for the second agent

u1::Int64: vertex for the first agent

u2::Int64: vertex for the second agent

MultiAgentPathFinding.LazyEdgeConflicts — Type

LazyEdgeConflicts

Lazy dict-like storage for the mapping (u, v) -> [(u, v)].

MultiAgentPathFinding.LazySwappingConflicts — Type

LazySwappingConflicts

Lazy dict-like storage for the mapping (u, v) -> [(v, u)].

MultiAgentPathFinding.LazyVertexConflicts — Type

LazyVertexConflicts

Lazy dict-like storage for the mapping v -> [v].

MultiAgentPathFinding.MAPF — Type

MAPF(
    g[, edge_weights];
    departures, arrivals[, departure_times, vertex_conflicts, edge_conflicts]
)

User-friendly constructor for a Multi-Agent Path Finding problem.

Departure times default to 1 for every agent, vertex conflicts default to LazyVertexConflicts and edge conflicts to LazySwappingConflicts.

MultiAgentPathFinding.MAPF — Type

struct MAPF{W<:Real, G<:Graphs.AbstractGraph{Int64}, M<:AbstractArray{W<:Real, 2}, VC, EC}

Instance of a Multi-Agent Path Finding problem with custom conflict rules.

Agents appear at their departure vertex when the departure time comes, and they disappear as soon as they have reached the arrival vertex.

Fields

g::Graphs.AbstractGraph{Int64}: underlying graph

edge_weights::AbstractMatrix{W} where W<:Real: matrix of edge weights

departures::Vector{Int64}: agent departure vertices

arrivals::Vector{Int64}: agent arrival vertices

departure_times::Vector{Int64}: agent departure times

vertex_conflicts::Any: dict-like object linking vertices to their incompatibility set

edge_conflicts::Any: dict-like object linking edges (as tuples) to their incompatibility set

MultiAgentPathFinding.Reservation — Type

mutable struct Reservation

Storage for vertices and edges that are known to be occupied.

Fields

forbidden_vertices::Set{Tuple{Int64, Int64}}

forbidden_edges::Set{Tuple{Int64, Int64, Int64}}

max_time::Int64: maximum time of all forbidden vertices (mutable)

MultiAgentPathFinding.Reservation — Method

Reservation()

Create an empty Reservation.

MultiAgentPathFinding.ShortestPathTree — Type

struct ShortestPathTree{T, W}

Storage for the result of Dijkstra's algorithm run backwards.

Fields

children::Vector: successor of each vertex in a shortest path

dists::Vector: distance of each vertex to the arrival

MultiAgentPathFinding.Solution — Type

Solution = Vector{TimedPath}

Vector of TimedPaths, one for each agent of a MAPF.

MultiAgentPathFinding.TimedPath — Type

struct TimedPath

Temporal path through a graph.

Fields

tdep::Int64: departure time

path::Vector{Int64}: sequence of vertices

MultiAgentPathFinding.TimedPath — Method

TimedPath(tdep)

Return an empty TimedPath.

MultiAgentPathFinding.all_non_empty — Method

all_non_empty(solution)

Check that all paths in a Solution are non empty.

MultiAgentPathFinding.arrival_time — Method

arrival_time(timed_path)

Return the departure time of a TimedPath plus its number of edges.

MultiAgentPathFinding.arrival_vertex — Method

arrival_vertex(timed_path)

Return the last vertex of a TimedPath.

MultiAgentPathFinding.backward_dijkstra — Method

backward_dijkstra(g, w; arr)

Run Dijkstra's algorithm backward from an arrival vertex, with specified edge weights. Return a ShortestPathTree where distances can be nothing.

MultiAgentPathFinding.build_path_astar — Method

build_path_astar(parents, arr, tarr)

Build a TimedPath from a dictionary of temporal parents, going backward from arr which was reached at tarr.

MultiAgentPathFinding.build_path_tree — Method

build_path_tree(spt, dep, arr, tdep)

Build a TimedPath from a ShortestPathTree, going from dep to arr and starting at time tdep.

MultiAgentPathFinding.compute_reservation — Function

compute_reservation(solution, mapf[; agents])

Compute a Reservation based on the vertices and edges occupied by solution (or a subset of its agents).

MultiAgentPathFinding.cooperative_astar — Function

cooperative_astar(solution, mapf, agents, spt_by_arr)

Create an empty Solution, a dictionary of ShortestPathTrees and apply cooperative_astar_from_trees!.

MultiAgentPathFinding.cooperative_astar_from_trees! — Method

cooperative_astar_from_trees!(solution, mapf, agents, spt_by_arr)

Modify a Solution by applying temporal_astar to a subset of agents while avoiding conflicts thanks to a Reservation.

Arguments

agents: subset of agents taken in order

spt_by_arr: dictionary of ShortestPathTrees, one for each arrival vertex

MultiAgentPathFinding.cooperative_astar_soft_from_trees! — Method

cooperative_astar_from_trees_soft!(solution, mapf, agents, spt_by_arr)

Does the same things as cooperative_astar_from_trees! but with temporal_astar_soft as a basic subroutine.

Arguments

agents, spt_by_arr: see cooperative_astar_from_trees!

conflict_price: see temporal_astar_soft

MultiAgentPathFinding.count_conflicts — Method

count_conflicts(a1, a2, solution, mapf[; tol=0])

Count the number of conflicts between agents a1 and a2 in a solution.

MultiAgentPathFinding.count_conflicts — Method

count_conflicts(solution, mapf[; tol=0])

Count the number of conflicts in a solution.

MultiAgentPathFinding.count_edge_conflicts — Method

count_edge_conflicts(a1, a2, solution, mapf[; tol=0])

Count the number of occurrences where the paths of a1 and a2 in the solution cross incompatible edges less than tol time steps apart.

MultiAgentPathFinding.count_vertex_conflicts — Method

count_vertex_conflicts(a1, a2, solution, mapf[; tol=0])

Count the number of occurrences where the paths of a1 and a2 in the solution visit incompatible vertices less than tol time steps apart.

MultiAgentPathFinding.departure_time — Method

departure_time(timed_path)

Return the departure time of a TimedPath.

MultiAgentPathFinding.departure_vertex — Method

departure_vertex(timed_path)

Return the first vertex of a TimedPath.

MultiAgentPathFinding.dijkstra_by_arrival — Method

dijkstra_by_arrival(mapf)

Run backward_dijkstra from each arrival vertex of a MAPF and return a dictionary of ShortestPathTrees.

MultiAgentPathFinding.double_search — Function

double_search(
    mapf, agents;
    feasibility_timeout, optimality_timeout, window,
    neighborhood_size, conflict_price, conflict_price_increase
)

Initialize a Solution with independent_dijkstra, then apply [feasibility_search!] to make it feasible, followed by optimality_search! to reduce its flowtime.

MultiAgentPathFinding.edge_at_time — Method

edge_at_time(timed_path, t)

Return the edge crossed by a TimedPath at a given time, throws an error if it does not exist.

MultiAgentPathFinding.empty_solution — Method

empty_solution(mapf)

Return a vector of empty TimedPaths.

MultiAgentPathFinding.exists_in_graph — Method

exists_in_graph(timed_path, g)

Check that a TimedPath is feasible in the graph g, i.e. that all vertices and edges exist.

MultiAgentPathFinding.feasibility_search! — Method

feasibility_search!(
    solution, mapf, spt_by_arr;
    feasibility_timeout, neighborhood_size, conflict_price, conflict_price_increase
)

Reduce the number of conflicts in an infeasible Solution with a variant of the MAPF-LNS2 algorithm from Li et al. (2022).

MultiAgentPathFinding.feasibility_search — Method

feasibility_search(
    mapf;
    feasibility_timeout, neighborhood_size, conflict_price, conflict_price_increase
)

Initialize a Solution with independent_dijkstra, and then apply feasibility_search! to reduce the number of conflicts.

MultiAgentPathFinding.find_conflict — Method

find_conflict(a1, a2, solution, mapf[; tol=0])

Find a conflict between agents a1 and a2 in a solution.

MultiAgentPathFinding.find_conflict — Method

find_conflict(solution, mapf[; tol=0])

Find a conflict in a solution.

MultiAgentPathFinding.find_edge_conflict — Method

find_edge_conflict(a1, a2, solution, mapf[; tol=0])

Find an occurrence where the paths of a1 and a2 in the solution cross incompatible edges less than tol time steps apart.

Return either a Conflict object or nothing.

MultiAgentPathFinding.find_vertex_conflict — Method

find_vertex_conflict(a1, a2, solution, mapf[; tol=0])

Find an occurrence where the paths of a1 and a2 in the solution visit incompatible vertices less than tol time steps apart.

Return either a Conflict object or nothing.

MultiAgentPathFinding.flowtime — Method

flowtime(solution, mapf)

Sum the flowtime of all the TimedPaths in a Solution.

MultiAgentPathFinding.independent_dijkstra — Method

independent_dijkstra(mapf)

Compute independent shortest paths for each agent.

MultiAgentPathFinding.independent_dijkstra_from_trees — Method

independent_dijkstra_from_trees(mapf, spt_by_arr)

Compute independent shortest paths for each agent based on the output of dijkstra_by_arrival (i.e. a dictionary of ShortestPathTrees)

MultiAgentPathFinding.is_collectively_feasible — Method

is_collectively_feasible(solution, mapf[; verbose])

Check whether a Solution contains any conflicts between agents.

MultiAgentPathFinding.is_feasible — Method

is_feasible(solution, mapf[; verbose])

Check whether a Solution is both individually and collectively feasible.

MultiAgentPathFinding.is_forbidden_edge — Method

is_forbidden_edge(reservation, t, u, v)

Check whether edge (u, v) is occupied at time t in a Reservation.

MultiAgentPathFinding.is_forbidden_vertex — Method

is_forbidden_vertex(reservation, t, v)

Check whether vertex v is occupied at time t in a Reservation.

MultiAgentPathFinding.is_individually_feasible — Method

is_individually_feasible(solution, mapf[; verbose])

Check whether a Solution is feasible when agents are considered separately.

MultiAgentPathFinding.makespan — Method

makespan(solution)

Compute the maximum arrival time of all the TimedPaths in a Solution.

MultiAgentPathFinding.max_time — Method

max_time(reservation)

Return the maximum time of all forbidden vertices in a Reservation.

MultiAgentPathFinding.nb_agents — Method

nb_agents(mapf)

Count the number of agents.

MultiAgentPathFinding.optimality_search — Function

optimality_search(
    mapf, agents, spt_by_arr;
    optimality_timeout, neighborhood_size
)

Initialize a Solution with cooperative_astar, and then apply optimality_search! to reduce its flowtime.

MultiAgentPathFinding.optimality_search! — Method

optimality_search!(
    solution, mapf, spt_by_arr;
    optimality_timeout, neighborhood_size
)

Reduce the flowtime of a feasible Solution with the MAPF-LNS algorithm from Li et al. (2021).

MultiAgentPathFinding.path_length_tree — Method

path_length_tree(spt, dep, arr)

Count the edges in a shortest path from dep to arr based on a ShortestPathTree.

MultiAgentPathFinding.path_weight — Method

path_weight(timed_path, mapf[; tmin, tmax])

Compute the weight of a TimedPath through a MAPF graph by summing edge weights between times tmin and tmax (which default to the departure_time and arrival_time).

MultiAgentPathFinding.random_neighborhood — Method

random_neighborhood(mapf, neighborhood_size)

Return a random subset of agents with a given size.

MultiAgentPathFinding.remove_agents! — Method

remove_agents!(solution, agents, mapf)

Remove a set of agents from a Solution and return a back up of their paths.

MultiAgentPathFinding.remove_arrival_vertex — Method

remove_arrival_vertex(timed_path)

Cut the last vertex from a TimedPath.

MultiAgentPathFinding.repeated_cooperative_astar — Method

repeated_cooperative_astar(mapf; coop_timeout)

Compute a dictionary of ShortestPathTrees with dijkstra_by_arrival, and then apply repeated_cooperative_astar_from_trees.

MultiAgentPathFinding.repeated_cooperative_astar_from_trees — Method

repeated_cooperative_astar_from_trees(mapf, spt_by_arr; coop_timeout)

Apply cooperative_astar_from_trees! repeatedly until a feasible solution is found or the timeout given by coop_timeout is reached.

MultiAgentPathFinding.select_agents — Method

select_agents(mapf, agents)

Select a subset of agents and return a new MAPF.

MultiAgentPathFinding.select_agents — Method

select_agents(mapf, A)

Select the first A agents and return a new MAPF.

MultiAgentPathFinding.temporal_astar — Method

temporal_astar(g, w; dep, arr, tdep, tmax, res, heuristic)

Apply temporal A* to a graph with specified edge weights. Subroutine of cooperative_astar_from_trees! and optimality_search!.

Keyword arguments

dep: departure vertex

arr: arrival vertex

tdep: departure time

tmax: max arrival time, after which the search stops and returns a partial path

res: reservation indicating occupied vertices and edges at various times

heuristic: indexable giving an underestimate of the remaining distance to arr

MultiAgentPathFinding.temporal_astar_soft — Method

temporal_astar_soft(g, w; dep, arr, tdep, tmax, res, heuristic, conflict_price)

Apply a bi-objective variant of temporal A* to a graph with specified edge weights. The objective is to minimize a combination of the number of conflicts and the path weight. Subroutine of feasibility_search!.

Keyword arguments

dep, arr, tdep, tmax, res, heuristic: see temporal_astar.

conflict_price: price given to the number of conflicts in the objective

MultiAgentPathFinding.update_reservation! — Method

update_reservation!(reservation, timed_path, mapf)

Add the vertices and edges occupied by a path to a Reservation.

MultiAgentPathFinding.vertex_at_time — Method

vertex_at_time(timed_path, t)

Return the vertex visited by a TimedPath at a given time, throws an error if it does not exist.

Index

MultiAgentPathFinding.MultiAgentPathFinding

MultiAgentPathFinding.Conflict

MultiAgentPathFinding.LazyEdgeConflicts

MultiAgentPathFinding.LazySwappingConflicts

MultiAgentPathFinding.LazyVertexConflicts

MultiAgentPathFinding.MAPF

MultiAgentPathFinding.MAPF

MultiAgentPathFinding.Reservation

MultiAgentPathFinding.Reservation

MultiAgentPathFinding.ShortestPathTree

MultiAgentPathFinding.Solution

MultiAgentPathFinding.TimedPath

MultiAgentPathFinding.TimedPath

MultiAgentPathFinding.all_non_empty

MultiAgentPathFinding.arrival_time

MultiAgentPathFinding.arrival_vertex

MultiAgentPathFinding.backward_dijkstra

MultiAgentPathFinding.build_path_astar

MultiAgentPathFinding.build_path_tree

MultiAgentPathFinding.compute_reservation

MultiAgentPathFinding.cooperative_astar

MultiAgentPathFinding.cooperative_astar_from_trees!

MultiAgentPathFinding.cooperative_astar_soft_from_trees!

MultiAgentPathFinding.count_conflicts

MultiAgentPathFinding.count_conflicts

MultiAgentPathFinding.count_edge_conflicts

MultiAgentPathFinding.count_vertex_conflicts

MultiAgentPathFinding.departure_time

MultiAgentPathFinding.departure_vertex

MultiAgentPathFinding.dijkstra_by_arrival

MultiAgentPathFinding.double_search

MultiAgentPathFinding.edge_at_time

MultiAgentPathFinding.empty_solution

MultiAgentPathFinding.exists_in_graph

MultiAgentPathFinding.feasibility_search

MultiAgentPathFinding.feasibility_search!

MultiAgentPathFinding.find_conflict

MultiAgentPathFinding.find_conflict

MultiAgentPathFinding.find_edge_conflict

MultiAgentPathFinding.find_vertex_conflict

MultiAgentPathFinding.flowtime

MultiAgentPathFinding.independent_dijkstra

MultiAgentPathFinding.independent_dijkstra_from_trees

MultiAgentPathFinding.is_collectively_feasible

MultiAgentPathFinding.is_feasible

MultiAgentPathFinding.is_forbidden_edge

MultiAgentPathFinding.is_forbidden_vertex

MultiAgentPathFinding.is_individually_feasible

MultiAgentPathFinding.makespan

MultiAgentPathFinding.max_time

MultiAgentPathFinding.nb_agents

MultiAgentPathFinding.optimality_search

MultiAgentPathFinding.optimality_search!

MultiAgentPathFinding.path_length_tree

MultiAgentPathFinding.path_weight

MultiAgentPathFinding.random_neighborhood

MultiAgentPathFinding.remove_agents!

MultiAgentPathFinding.remove_arrival_vertex

MultiAgentPathFinding.repeated_cooperative_astar

MultiAgentPathFinding.repeated_cooperative_astar_from_trees

MultiAgentPathFinding.select_agents

MultiAgentPathFinding.select_agents

MultiAgentPathFinding.temporal_astar

MultiAgentPathFinding.temporal_astar_soft

MultiAgentPathFinding.update_reservation!

MultiAgentPathFinding.vertex_at_time