API reference

Types

MultiAgentPathFinding.MAPF — Type

MAPF{G}

Instance of a Multi-Agent PathFinding problem with custom conflict rules.

Fields

g::G

edge_indices::Dict{Tuple{Int,Int},Int}

edge_colptr::Vector{Int}

edge_rowval::Vector{Int}

edge_weights_vec::Vector{Float64}

vertex_conflicts::Vector{Vector{Int}}

edge_conflicts::Vector{Vector{Int}}

sources::Vector{Int}

destinations::Vector{Int}

starting_times::Vector{Int}

MultiAgentPathFinding.build_weights_matrix — Function

build_edge_weights_matrix(mapf[, edge_weights_vec])

Turn a vector edge_weights_vec into a sparse weighted adjacency matrix for the graph mapf.g.

This function doesn't allocate because the necessary index information is already in the MAPF object.

MultiAgentPathFinding.nb_agents — Method

nb_agents(mapf)

Count the number of agents in mapf.

MultiAgentPathFinding.TimedPath — Type

TimedPath

Temporal path through a graph.

Fields

t0::Int

path::Vector{Int}

MultiAgentPathFinding.path_to_vec — Method

path_to_vec(timed_path, mapf)

Encode a timed_path as a dense integer vector that counts edge crossings.

MultiAgentPathFinding.path_to_vec_sparse — Method

path_to_vec_sparse(timed_path, mapf)

Encode a timed_path as a sparse integer vector that counts edge crossings.

MultiAgentPathFinding.Reservation — Type

Reservation

Storage for vertices and edges that are already occupied.

Fields

forbidden_vertices::Set{Tuple{Int,Int}}: set of tuples (t,v)

forbidden_edges::Set{Tuple{Int,Int}}: set of tuples (t,e)

MultiAgentPathFinding.compute_reservation — Method

compute_reservation(solution, mapf; [agents])

Compute a Reservation based on the vertices and edges occupied by solution (or a subset of its agents).

MultiAgentPathFinding.is_forbidden_edge — Method

is_forbidden_edge(reservation, t, e)

Check whether edge e is occupied at time t.

MultiAgentPathFinding.is_forbidden_vertex — Method

is_forbidden_vertex(reservation, t, v)

Check whether vertex v is occupied at time t.

MultiAgentPathFinding.update_reservation! — Method

update_reservation!(reservation, timed_path, mapf)

Add the vertices and edges occupied by timed_path to reservation.

MultiAgentPathFinding.Solution — Type

Solution = Vector{TimedPath}

Vector of TimedPaths, one for each agent of a MAPF.

MultiAgentPathFinding.remove_agents! — Method

remove_agents!(solution, agents, mapf)

Remove a set of agents from a solution and return a back up of their paths.

Benchmark

MultiAgentPathFinding.BenchmarkMAPF — Type

BenchmarkMAPF = MAPF{SparseGridGraph{Int64,Float64}}

Concrete subtype of MAPF designed for the benchmark instances of Sturtevant.

MultiAgentPathFinding.benchmark_mapf — Method

benchmark_mapf(map_path, scenario_path[; buckets])

MultiAgentPathFinding.is_solvable — Method

is_solvable(mapf::BenchmarkMAPF)

MultiAgentPathFinding.display_benchmark_map — Method

display_benchmark_mapf(map_matrix::Matrix{Char})

MultiAgentPathFinding.BenchmarkProblem — Type

BenchmarkProblem

Fields

index::Int

bucket::Int

map::String

width::Int

height::Int

start_i::Int

start_j::Int

goal_i::Int

goal_j::Int

optimal_length::Float64

MultiAgentPathFinding.read_benchmark_map — Method

read_benchmark_map(map_path)

MultiAgentPathFinding.read_benchmark_scenario — Method

read_benchmark_scenario(scenario_path, map_path)

Shortest paths

MultiAgentPathFinding.ShortestPathTree — Type

ShortestPathTree{T,W}

Storage for the result of Dijkstra's algorithm.

Fields

forward::Bool

parents::Vector{T}

dists::Vector{W}

MultiAgentPathFinding.backward_dijkstra — Method

backward_dijkstra(g, d, w)

MultiAgentPathFinding.build_dijkstra_path — Method

build_dijkstra_path(shortest_path_tree, t0, s, d)

MultiAgentPathFinding.forward_dijkstra — Method

forward_dijkstra(g, s, w)

MultiAgentPathFinding.build_astar_path — Method

build_astar_path(parents, t0, s, tf, d)

MultiAgentPathFinding.temporal_astar — Method

temporal_astar(g, s, d, t0, w[, reservation; heuristic, conflict_price])

MultiAgentPathFinding.temporal_astar_hard — Method

temporal_astar_hard(g, s, d, t0, w[, reservation; heuristic])

MultiAgentPathFinding.temporal_astar_soft — Method

temporal_astar_soft(g, s, d, t0, w[, reservation; heuristic, conflict_price])

Local search

Solution evaluation

MultiAgentPathFinding.flowtime — Function

flowtime(solution, mapf[, edge_weights_vec])

MultiAgentPathFinding.max_time — Method

max_time(solution)

MultiAgentPathFinding.path_weight — Method

path_weight(timed_path, mapf[, edge_weights_vec])

MultiAgentPathFinding.has_empty_paths — Method

has_empty_paths(solution)

MultiAgentPathFinding.is_feasible — Method

is_feasible(solution)

Machine learning

Index

MultiAgentPathFinding.BenchmarkMAPF

MultiAgentPathFinding.BenchmarkProblem

MultiAgentPathFinding.MAPF

MultiAgentPathFinding.Reservation

MultiAgentPathFinding.ShortestPathTree

MultiAgentPathFinding.Solution

MultiAgentPathFinding.TimedPath

MultiAgentPathFinding.backward_dijkstra

MultiAgentPathFinding.benchmark_mapf

MultiAgentPathFinding.build_astar_path

MultiAgentPathFinding.build_dijkstra_path

MultiAgentPathFinding.build_weights_matrix

MultiAgentPathFinding.compute_reservation

MultiAgentPathFinding.display_benchmark_map

MultiAgentPathFinding.flowtime

MultiAgentPathFinding.forward_dijkstra

MultiAgentPathFinding.has_empty_paths

MultiAgentPathFinding.is_feasible

MultiAgentPathFinding.is_forbidden_edge

MultiAgentPathFinding.is_forbidden_vertex

MultiAgentPathFinding.is_solvable

MultiAgentPathFinding.max_time

MultiAgentPathFinding.nb_agents

MultiAgentPathFinding.path_to_vec

MultiAgentPathFinding.path_to_vec_sparse

MultiAgentPathFinding.path_weight

MultiAgentPathFinding.read_benchmark_map

MultiAgentPathFinding.read_benchmark_scenario

MultiAgentPathFinding.remove_agents!

MultiAgentPathFinding.temporal_astar

MultiAgentPathFinding.temporal_astar_hard

MultiAgentPathFinding.temporal_astar_soft

MultiAgentPathFinding.update_reservation!