API reference

Only exported names are part of the API.

MultiAgentPathFinding — Module

MultiAgentPathFinding

A package for Multi-Agent Path Finding instances and algorithms.

Exports

Conflict

MAPF

Solution

TimedPath

cooperative_astar

double_search

feasibility_search

find_conflict

independent_dijkstra

is_feasible

list_map_names

list_scenario_names

nb_agents

optimality_search

path_cost

read_benchmark

select_agents

total_path_cost

Structures

MultiAgentPathFinding.MAPF — Type

struct MAPF{W<:Real, G<:Graphs.AbstractGraph{Int64}, M<:AbstractArray{W<:Real, 2}, VC, EC}

Instance of a Multi-Agent Path Finding problem with custom conflict rules.

Agents appear at their departure vertex when the departure time comes, and they disappear as soon as they have reached the arrival vertex.

Fields

g::Graphs.AbstractGraph{Int64}: underlying graph

edge_weights::AbstractMatrix{W} where W<:Real: matrix of edge weights

departures::Vector{Int64}: agent departure vertices

arrivals::Vector{Int64}: agent arrival vertices

departure_times::Vector{Int64}: agent departure times

vertex_conflicts::Any: dict-like object linking vertices to their incompatibility set

edge_conflicts::Any: dict-like object linking edges (as tuples) to their incompatibility set

MultiAgentPathFinding.TimedPath — Type

struct TimedPath

Temporal path through a graph.

Fields

tdep::Int64: departure time

path::Vector{Int64}: sequence of vertices

MultiAgentPathFinding.Solution — Type

Solution = Vector{TimedPath}

Vector of TimedPaths, one for each agent of a MAPF.

MultiAgentPathFinding.Conflict — Type

struct Conflict

Store a conflict between two agents for debugging purposes.

Fields

name::Symbol: type of conflict (:vertex or :edge)

a1::Int64: first agent

a2::Int64: second agent

t1::Int64: time for the first agent

t2::Int64: time for the second agent

u1::Int64: vertex for the first agent

u2::Int64: vertex for the second agent

Access

MultiAgentPathFinding.nb_agents — Function

nb_agents(mapf::MAPF) -> Int64

Count the number of agents.

MultiAgentPathFinding.select_agents — Function

select_agents(
    mapf::MAPF,
    agents::AbstractVector{<:Integer}
) -> MAPF

Select a subset of agents and return a new MAPF.

Feasibility and cost

MultiAgentPathFinding.path_cost — Function

path_cost(
    timed_path::TimedPath,
    mapf::MAPF{W, G, M} where {G<:Graphs.AbstractGraph{Int64}, M<:AbstractArray{W, 2}};
    tmin,
    tmax
) -> Any

Compute the weight of a timed path by summing edge weights between times tmin and tmax (which default to the departure and arrival time).

MultiAgentPathFinding.total_path_cost — Function

total_path_cost(
    solution::Vector{TimedPath},
    mapf::MAPF
) -> Any

Sum the weight of all the paths in a solution.

MultiAgentPathFinding.is_feasible — Function

is_feasible(
    solution::Vector{TimedPath},
    mapf::MAPF;
    verbose
) -> Bool

Check whether a solution is both individually and collectively feasible (correct paths and no conflicts).

MultiAgentPathFinding.find_conflict — Function

find_conflict(
    solution::Vector{TimedPath},
    mapf::MAPF;
    tol
) -> Union{Nothing, Conflict}

Find a conflict in a solution.

find_conflict(
    a1::Integer,
    a2::Integer,
    solution::Vector{TimedPath},
    mapf::MAPF;
    tol
) -> Union{Nothing, Conflict}

Find a conflict between agents a1 and a2 in a solution.

Basic algorithms

MultiAgentPathFinding.independent_dijkstra — Function

independent_dijkstra(
    mapf::MAPF;
    show_progress
) -> Vector{TimedPath}

Compute independent shortest paths for each agent.

Returns a Solution.

MultiAgentPathFinding.cooperative_astar — Function

cooperative_astar(mapf::MAPF; ...) -> Vector
cooperative_astar(
    mapf::MAPF,
    agents::AbstractVector{<:Integer};
    show_progress
) -> Vector

Solve a MAPF problem with the cooperative A* algorithm of Silver (2005), see https://ojs.aaai.org/index.php/AIIDE/article/view/18726.

Returns a Solution.

Local search

MultiAgentPathFinding.feasibility_search — Function

feasibility_search(
    mapf::MAPF;
    feasibility_timeout,
    neighborhood_size,
    conflict_price,
    conflict_price_increase,
    show_progress
) -> Tuple{Vector{TimedPath}, Dict}

Run independent_dijkstra and then reduce the number of conflicts with a variant of the MAPF-LNS2 algorithm from Li et al. (2022), see https://ojs.aaai.org/index.php/AAAI/article/view/21266.

Returns a tuple containing a Solution and a dictionary of statistics.

MultiAgentPathFinding.optimality_search — Function

optimality_search(mapf::MAPF; ...) -> Tuple{Vector, Dict}
optimality_search(
    mapf::MAPF,
    agents;
    optimality_timeout,
    neighborhood_size,
    show_progress
) -> Tuple{Vector, Dict}

Run cooperative_astar and then reduce the total path cost with the MAPF-LNS algorithm from Li et al. (2021), see https://www.ijcai.org/proceedings/2021/568.

Returns a tuple containing a Solution and a dictionary of statistics.

MultiAgentPathFinding.double_search — Function

double_search(mapf::MAPF; ...) -> Tuple{Vector, Dict}
double_search(
    mapf::MAPF,
    agents;
    feasibility_timeout,
    optimality_timeout,
    neighborhood_size,
    conflict_price,
    conflict_price_increase,
    show_progress
) -> Tuple{Vector, Dict}

Combine feasibility_search and optimality_search, see https://pastel.hal.science/tel-04053322.

Returns a tuple containing a Solution and a dictionary of statistics.

Benchmarks

MultiAgentPathFinding.read_benchmark — Function

read_benchmark(
    map_name::AbstractString,
    scenario_name::AbstractString;
    check
) -> MAPF{Float64, SimpleWeightedGraphs.SimpleWeightedGraph{Int64, Float64}, SparseArrays.SparseMatrixCSC{Float64, Int64}, MultiAgentPathFinding.LazyVertexConflicts, MultiAgentPathFinding.LazySwappingConflicts}

Create a MAPF instance by reading a map ("something.map") and scenario ("something.scen") from files.

See possible names at https://movingai.com/benchmarks/mapf/index.html (data will be downloaded automatically).

MultiAgentPathFinding.list_map_names — Function

list_map_names() -> Vector{String}

List available maps from the benchmark set.

MultiAgentPathFinding.list_scenario_names — Function

list_scenario_names() -> Vector{String}

List available scenarios from the benchmark set.

Internals

MultiAgentPathFinding.LazyEdgeConflicts — Type

struct LazyEdgeConflicts

Lazy dict-like storage for the mapping (u, v) -> [(u, v)].

MultiAgentPathFinding.LazySwappingConflicts — Type

struct LazySwappingConflicts

Lazy dict-like storage for the mapping (u, v) -> [(v, u)].

MultiAgentPathFinding.LazyVertexConflicts — Type

struct LazyVertexConflicts

Lazy dict-like storage for the mapping v -> [v].

MultiAgentPathFinding.MAPFBenchmarkProblem — Type

struct MAPFBenchmarkProblem

Encode one agent of a MAPF scenario.

Fields

index::Int64

bucket::Int64

map_path::String

width::Int64

height::Int64

start_i::Int64

start_j::Int64

goal_i::Int64

goal_j::Int64

optimal_length::Float64

MultiAgentPathFinding.Reservation — Type

mutable struct Reservation

Storage for vertices and edges that are known to be occupied.

Fields

forbidden_vertices::Set{Tuple{Int64, Int64}}

forbidden_edges::Set{Tuple{Int64, Int64, Int64}}

max_time::Int64: maximum time of all forbidden vertices (mutable)

MultiAgentPathFinding.Reservation — Method

Reservation() -> MultiAgentPathFinding.Reservation

Create an empty reservation.

MultiAgentPathFinding.ShortestPathTree — Type

struct ShortestPathTree{T, W}

Storage for the result of Dijkstra's algorithm run backwards.

Fields

children::Vector: successor of each vertex in a shortest path

dists::Vector: distance of each vertex to the arrival

MultiAgentPathFinding.all_non_empty — Method

all_non_empty(solution::Vector{TimedPath}) -> Bool

Check that all paths in a solution are non empty.

MultiAgentPathFinding.arrival_time — Method

arrival_time(timed_path::TimedPath) -> Int64

Return the departure time of a timed path plus its number of edges.

MultiAgentPathFinding.arrival_vertex — Method

arrival_vertex(timed_path::TimedPath) -> Int64

Return the last vertex of a timed path.

MultiAgentPathFinding.backward_dijkstra — Method

backward_dijkstra(
    g::Graphs.AbstractGraph{T},
    w::AbstractArray{W, 2};
    arr
)

Run Dijkstra's algorithm backward from an arrival vertex, with specified edge weights.

Returns a ShortestPathTree where distances can be nothing.

MultiAgentPathFinding.benchmark_cell_color — Method

benchmark_cell_color(
    c::Char
) -> ColorTypes.RGB{FixedPointNumbers.N0f8}

Give a color object corresponding to the type of cell.

To visualize a map in VSCode, just run cell_color.(map_matrix) in the REPL.

MultiAgentPathFinding.build_path_astar — Method

build_path_astar(
    parents::Dict,
    arr::Integer,
    tarr::Integer
) -> TimedPath

Build a TimedPath from a dictionary of temporal parents, going backward from arr which was reached at tarr.

MultiAgentPathFinding.build_path_tree — Method

build_path_tree(
    spt::MultiAgentPathFinding.ShortestPathTree{T},
    dep::Integer,
    arr::Integer,
    tdep::Integer
) -> TimedPath

Build a TimedPath from a ShortestPathTree, going from dep to arr and starting at time tdep.

MultiAgentPathFinding.compute_reservation — Function

compute_reservation(
    solution::Vector{TimedPath},
    mapf::MAPF
) -> MultiAgentPathFinding.Reservation
compute_reservation(
    solution::Vector{TimedPath},
    mapf::MAPF,
    agents::AbstractVector{<:Integer}
) -> MultiAgentPathFinding.Reservation

Compute a Reservation based on the vertices and edges occupied by a solution (or a subset of its agents).

MultiAgentPathFinding.count_conflicts — Method

count_conflicts(
    a1::Integer,
    a2::Integer,
    solution::Vector{TimedPath},
    mapf::MAPF;
    tol
) -> Any

Count the number of conflicts between agents a1 and a2 in a solution.

MultiAgentPathFinding.count_conflicts — Method

count_conflicts(
    solution::Vector{TimedPath},
    mapf::MAPF;
    tol
) -> Any

Count the number of conflicts in a solution.

MultiAgentPathFinding.count_edge_conflicts — Method

count_edge_conflicts(
    a1::Integer,
    a2::Integer,
    solution::Vector{TimedPath},
    mapf::MAPF;
    tol
) -> Any

Count the number of occurrences where the paths of a1 and a2 in the solution cross incompatible edges less than tol time steps apart.

MultiAgentPathFinding.count_vertex_conflicts — Method

count_vertex_conflicts(
    a1::Integer,
    a2::Integer,
    solution::Vector{TimedPath},
    mapf::MAPF;
    tol
) -> Any

Count the number of occurrences where the paths of a1 and a2 in the solution visit incompatible vertices less than tol time steps apart.

MultiAgentPathFinding.departure_time — Method

departure_time(timed_path::TimedPath) -> Int64

Return the departure time of a timed path.

MultiAgentPathFinding.departure_vertex — Method

departure_vertex(timed_path::TimedPath) -> Int64

Return the first vertex of a timed path.

MultiAgentPathFinding.dijkstra_by_arrival — Method

dijkstra_by_arrival(
    mapf::MAPF{W, G, M} where {G<:Graphs.AbstractGraph{Int64}, M<:AbstractArray{W, 2}};
    show_progress
) -> Dict{Int64, V} where V<:(MultiAgentPathFinding.ShortestPathTree{Int64})

Run backward_dijkstra from each arrival vertex of a MAPF.

Returns a dictionary of ShortestPathTrees.

MultiAgentPathFinding.edge_at_time — Method

edge_at_time(
    timed_path::TimedPath,
    t::Integer
) -> Tuple{Any, Any}

Return the edge crossed by a timed path at a given time, throws an error if it does not exist.

MultiAgentPathFinding.empty_solution — Method

empty_solution(mapf::MAPF) -> Vector

Return a vector of empty TimedPaths.

MultiAgentPathFinding.exists_in_graph — Method

exists_in_graph(
    timed_path::TimedPath,
    g::Graphs.AbstractGraph
) -> Bool

Check that a timed path is feasible in the graph g, i.e. that all vertices and edges exist.

MultiAgentPathFinding.find_edge_conflict — Method

find_edge_conflict(
    a1::Integer,
    a2::Integer,
    solution::Vector{TimedPath},
    mapf::MAPF;
    tol
) -> Union{Nothing, Conflict}

Find an occurrence where the paths of a1 and a2 in the solution cross incompatible edges less than tol time steps apart.

Returns either a Conflict object or nothing.

MultiAgentPathFinding.find_vertex_conflict — Method

find_vertex_conflict(
    a1::Integer,
    a2::Integer,
    solution::Vector{TimedPath},
    mapf::MAPF;
    tol
) -> Union{Nothing, Conflict}

Find an occurrence where the paths of a1 and a2 in the solution visit incompatible vertices less than tol time steps apart.

Returns either a Conflict object or nothing.

MultiAgentPathFinding.independent_dijkstra_from_trees — Method

independent_dijkstra_from_trees(
    mapf::MAPF,
    spt_by_arr::Dict{<:Integer, <:MultiAgentPathFinding.ShortestPathTree}
) -> Vector{TimedPath}

Compute independent shortest paths for each agent based on the output of dijkstra_by_arrival (i.e. a dictionary of ShortestPathTrees)

MultiAgentPathFinding.is_collectively_feasible — Method

is_collectively_feasible(
    solution::Vector{TimedPath},
    mapf::MAPF;
    verbose
) -> Bool

Check whether a solution contains any conflicts between agents.

MultiAgentPathFinding.is_forbidden_edge — Method

is_forbidden_edge(
    reservation::MultiAgentPathFinding.Reservation,
    t::Integer,
    u::Integer,
    v::Integer
) -> Bool

Check whether edge (u, v) is occupied at time t in a reservation.

MultiAgentPathFinding.is_forbidden_vertex — Method

is_forbidden_vertex(
    reservation::MultiAgentPathFinding.Reservation,
    t::Integer,
    v::Integer
) -> Bool

Check whether vertex v is occupied at time t in a reservation.

MultiAgentPathFinding.is_individually_feasible — Method

is_individually_feasible(
    solution::Vector{TimedPath},
    mapf::MAPF;
    verbose
) -> Bool

Check whether a solution is feasible when agents are considered separately.

MultiAgentPathFinding.makespan — Method

makespan(solution::Vector{TimedPath}) -> Int64

Compute the maximum arrival time of all the paths in a solution.

MultiAgentPathFinding.map_from_scenario — Method

map_from_scenario(scenario_name::AbstractString) -> String

Return the map associated with a benchmark scenario.

MultiAgentPathFinding.max_time — Method

max_time(
    reservation::MultiAgentPathFinding.Reservation
) -> Int64

Return the maximum time of all forbidden vertices in a reservation.

MultiAgentPathFinding.parse_benchmark_map — Method

parse_benchmark_map(
    map_matrix::Matrix{Char}
) -> Tuple{SimpleWeightedGraphs.SimpleWeightedGraph{Int64, Float64}, Dict{Tuple{Int64, Int64}, Int64}}

Create a sparse grid graph from a map specified as a matrix of characters.

MultiAgentPathFinding.parse_benchmark_scenario — Method

parse_benchmark_scenario(
    scenario::Vector{MultiAgentPathFinding.MAPFBenchmarkProblem},
    coord_to_vertex::Dict
) -> Tuple{Vector{Int64}, Vector{Int64}}

Turn a scenario into vectors of departure and arrival vertices.

MultiAgentPathFinding.path_length_tree — Method

path_length_tree(
    spt::MultiAgentPathFinding.ShortestPathTree,
    dep::Integer,
    arr::Integer
) -> Int64

Count the edges in a shortest path from dep to arr based on a ShortestPathTree.

MultiAgentPathFinding.random_neighborhood — Method

random_neighborhood(
    mapf::MAPF,
    neighborhood_size::Integer
) -> AbstractArray

Return a random subset of agents with a given size.

MultiAgentPathFinding.read_benchmark_map — Method

read_benchmark_map(map_name::AbstractString) -> Matrix{Char}

Read a map matrix from a text file.

Returns a Matrix{Char}.

MultiAgentPathFinding.read_benchmark_scenario — Method

read_benchmark_scenario(
    scenario_name::AbstractString,
    map_name::AbstractString
) -> Vector{MultiAgentPathFinding.MAPFBenchmarkProblem}

Read a scenario from a text file, and check that it corresponds to a given map.

Returns a Vector{MAPFBenchmarkProblem}.

MultiAgentPathFinding.remove_agents! — Method

remove_agents!(
    solution::Vector{TimedPath},
    agents::AbstractVector{<:Integer},
    mapf::MAPF
) -> Dict

Remove a set of agents from a solution and return a back up of their paths.

MultiAgentPathFinding.scenarios_from_map — Method

scenarios_from_map(
    map_name::AbstractString
) -> Vector{String}

List the scenarios associated with a benchmark map.

MultiAgentPathFinding.temporal_astar — Method

temporal_astar(
    g::Graphs.AbstractGraph{V},
    w::AbstractArray{W, 2};
    dep,
    arr,
    tdep,
    res,
    heuristic
)

Apply temporal A* to a graph with specified edge weights.

Keyword arguments

dep: departure vertex

arr: arrival vertex

tdep: departure time

res: reservation indicating occupied vertices and edges at various times

heuristic: indexable giving an underestimate of the remaining distance to arr

MultiAgentPathFinding.temporal_astar_soft — Method

temporal_astar_soft(
    g::Graphs.AbstractGraph{V},
    w::AbstractArray{W, 2};
    dep,
    arr,
    tdep,
    res,
    heuristic,
    conflict_price
)

Apply a bi-objective variant of temporal A* to a graph with specified edge weights. The objective is to minimize a combination of the number of conflicts and the path weight.

Keyword arguments

dep, arr, tdep, res, heuristic: see temporal_astar.

conflict_price: price given to the number of conflicts in the objective

MultiAgentPathFinding.update_reservation! — Method

update_reservation!(
    reservation::MultiAgentPathFinding.Reservation,
    timed_path::TimedPath,
    mapf::MAPF
)

Add the vertices and edges occupied by a timed path to a reservation.

MultiAgentPathFinding.vertex_at_time — Method

vertex_at_time(timed_path::TimedPath, t::Integer) -> Any

Return the vertex visited by a timed path at a given time, throws an error if it does not exist.

Index

MultiAgentPathFinding

MultiAgentPathFinding.Conflict

MultiAgentPathFinding.LazyEdgeConflicts

MultiAgentPathFinding.LazySwappingConflicts

MultiAgentPathFinding.LazyVertexConflicts

MultiAgentPathFinding.MAPF

MultiAgentPathFinding.MAPFBenchmarkProblem

MultiAgentPathFinding.Reservation

MultiAgentPathFinding.Reservation

MultiAgentPathFinding.ShortestPathTree

MultiAgentPathFinding.Solution

MultiAgentPathFinding.TimedPath

MultiAgentPathFinding.all_non_empty

MultiAgentPathFinding.arrival_time

MultiAgentPathFinding.arrival_vertex

MultiAgentPathFinding.backward_dijkstra

MultiAgentPathFinding.benchmark_cell_color

MultiAgentPathFinding.build_path_astar

MultiAgentPathFinding.build_path_tree

MultiAgentPathFinding.compute_reservation

MultiAgentPathFinding.cooperative_astar

MultiAgentPathFinding.count_conflicts

MultiAgentPathFinding.count_conflicts

MultiAgentPathFinding.count_edge_conflicts

MultiAgentPathFinding.count_vertex_conflicts

MultiAgentPathFinding.departure_time

MultiAgentPathFinding.departure_vertex

MultiAgentPathFinding.dijkstra_by_arrival

MultiAgentPathFinding.double_search

MultiAgentPathFinding.edge_at_time

MultiAgentPathFinding.empty_solution

MultiAgentPathFinding.exists_in_graph

MultiAgentPathFinding.feasibility_search

MultiAgentPathFinding.find_conflict

MultiAgentPathFinding.find_edge_conflict

MultiAgentPathFinding.find_vertex_conflict

MultiAgentPathFinding.independent_dijkstra

MultiAgentPathFinding.independent_dijkstra_from_trees

MultiAgentPathFinding.is_collectively_feasible

MultiAgentPathFinding.is_feasible

MultiAgentPathFinding.is_forbidden_edge

MultiAgentPathFinding.is_forbidden_vertex

MultiAgentPathFinding.is_individually_feasible

MultiAgentPathFinding.list_map_names

MultiAgentPathFinding.list_scenario_names

MultiAgentPathFinding.makespan

MultiAgentPathFinding.map_from_scenario

MultiAgentPathFinding.max_time

MultiAgentPathFinding.nb_agents

MultiAgentPathFinding.optimality_search

MultiAgentPathFinding.parse_benchmark_map

MultiAgentPathFinding.parse_benchmark_scenario

MultiAgentPathFinding.path_cost

MultiAgentPathFinding.path_length_tree

MultiAgentPathFinding.random_neighborhood

MultiAgentPathFinding.read_benchmark

MultiAgentPathFinding.read_benchmark_map

MultiAgentPathFinding.read_benchmark_scenario

MultiAgentPathFinding.remove_agents!

MultiAgentPathFinding.scenarios_from_map

MultiAgentPathFinding.select_agents

MultiAgentPathFinding.temporal_astar

MultiAgentPathFinding.temporal_astar_soft

MultiAgentPathFinding.total_path_cost

MultiAgentPathFinding.update_reservation!

MultiAgentPathFinding.vertex_at_time